requirements.txt
bottle
requests
bottle-beaker